refactor(frontend): rename file to `Schemas.tsx`
